Adding eset and dset flags; restoring domain events
Altering default configuration from the command line
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Added eset and dset flags to override Parameter database settings in the
config file for ECJ and the domain model, respectively.  The overrides
to ECJ's parameter database are made in the EPSimpleEvolutionState
class; the overrides to the domain model configuration are prepared in
the Problem's prepareEvaluator method and actually overriden after model
creation in the OligopolyProblem code.

As a precaution against configuration loss, the actual configuration
used by both components needs to be stored, either to the standard
output as part of a run log or to a dedicated configuration log.  Right
now this is done by displaying the command line arguments to stderr when
collecting eset and dset changes.

In order to get the domain model configured correctly via dset,
initialization needs to be delayed.  The constructor to
OligopolySimulation now just initializes a bare-bones (seed only)
version of MarketSimulation and loads and registers the configuration
file.  A separate initialize() step is called prior to model execution. 
This allows for the OligopolyProblem to modify the model's configuration
files prior to initialization.  Some final attributes needed to be
removed from some members since their initialization is no longer done
in the constructor.


Adding Domain Event Handling Back
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The hooks to set domain model events was temporarily removed.  I have
just added it back.  There at least two places where domain events
descriptions could reside prior to being placed into the domain model:
either in the problem instance or the EvolutionState.  I chose to go
with the latter again since I didn't want a static member handling
shared information between problems, worrying about concurrency since
Problem evaluation is multi-threaded.  The main evolution loop where the
other ECJ events are hooked in is part of the single-threaded portion of
the ECJ code.  So, as before, domain events are created by triggering a
specially created EventProcedure that fills in the domain_events member
of EPSimpleEvolutionState.  The problems use this array to configure the
domain model.

ECJEventProcedureManager now has a toString() method to display the
current state of its queues.