Adding spatial primary subpop eval grouper; rename existing non-spatial version.
The new SpatialPrimaryFSEvalGrouper will only group Individuals in the
primary subpopulation together if they are local to one another via the
Space interface.  Subpopulations that do not implement Space will throw
a runtime error if they are used as primary_subpopulations.  Individuals
outside of the primary subpopulation are still chosen at random.