Adding FileManager; updating reporting procedures
There are two ways I can think of to deal with reporting: take a static
approach like log4j or pass around file handles via a file manager. 
Since it is easy to implement, I went ahead and added in a
FileManager to EPSimpleEvolutionState that gets shared with the domain
problems.  Procedures simply request a file by path name (or with
initialization arguments); if the file is already open then the file is
returned or a new file is created / exception thrown if the file cannot
be created.  This make the use of files pretty easy and provides a
single way of retrieving the file regardless if the file is open,
exists, or not.

DelimitedOutFile out = state.file_manager.getDelimitedOutFile(path,
format)

OligopolySimulation gets its file manager set by OligopolyProblem prior
to model initialization.  The Procedures that perform output access the
file manager via the OligopolySimulation object that is passed while
processing an event and its context.

The main issue (though it's not horrible by any stretch) is that files,
once open, need to remain open until the end of the simulation.  This is
because there are no policies in place to decide whether or not a file
should be overwritten or appended.  Instead, the default is that if a
file exists prior to its first opening, the file is replaced.  That file
then remains open and accessible through the file manager until the end
of the experiment to prevent overwriting.