Added fitness to EvaluationGroup for statistics collection.
FileManager
-----------
Make sure that active_files are not stored on serialization.  The member
is now marked as transient and initialized to empty on deserialization.

EvaluationGroup
---------------
Fitness values are now stored with the group.  These fitnesses are set
by the OligopolyProblem but are not used during selection.  Their
primary purpose is to allow fitnesses that won't be used for selection
to be passed back from the Problems for data collection.

OligopolyProblem
----------------
Modified to set fitnesses in EvaluationGroup, not just with the
individual.

MASProblem
----------
Added the ability to clone the problem.