Bug fix time display in GroupedEvaluator; modified command line setter; created our own Evolve
In GroupedEvaluator, the evaluation time display didn't set the average
time correctly; I fixed that.  In addition I added an average thread
time to get the average amount of time each thread spends on
calculations to track performance differences in the future.

The command line -eset command previous resided in
EPSimpleEvolutionState.  But, the random number seeds are set in Evolve.
In order to get the seeds set correctly, the parameter database needs to
be updated early in the initialization process.  So, I added a modified
version of Evolve to Agency to include the parameter database process.

EPSimpleEvolutionState was modified to reflect these changes as well; it
still sets the domain model overrides, however.