Fix missing trigger for ATEND in EPSimpleEvolutionState
EC and Domain interval events that should be triggered using the
keyword "end" weren't being triggered on the EC side.  This was
because Interval.ATEND was never passed as a generational context
in EPSimpleEvolutionState.  All domain events that were to be triggered
in the last generation also didn't get triggered because they rely
upon the EC event system to create them.  

EPSimpleEvolutionState now uses a generation_cxt variable in the main
evolve() loop to store the right context to send to the EventManager.