Add EchoNodeGP to silence warnings.
ECJ's tree building algorithms require all types to have a terminal
defined for them.  This creates a problem, which up until now has just
been an ignored set of warnings, because some types require at least two
levels in order to return any type of value (e.g. an individual's root
ruleset or a root node for a particular rule).  

To silence these warnings and void possible problems, I created a new
terminal called EchoGP which does nothing.  That is, it does not alter
the input it receives and just serves as a place holder.  Now, rule and
ruleset nodes that are used as terminals (which shouldn't be the case
ever) will simply do nothing by using instances of EchoGP.

The Oligopoly/short-test has been updated to reflect this change. 
Currently, the program is running without any warning messages.