EC behavior fixes; Firm setup fixes
I've modified what I've done for EC agents to rely more exclusively on
the action system, as well as simplifying how prices and production
levels are set.

SetPriceAction now incudes a non-runtime error reject() method that does
nothing.

Firms now have last_production set by the simulation.  For oligopoly,
this is done in setupFirm().  An associated setLastProduction() method
is now included in Firm.

Firm production action verification now rejects negative quantities.

The way the EC rules interact with the agents has been changed.  Now,
rules trigger a set of scalings on a good:price or good:production
mapping, initially set by the last production and last price for a good.
 Only after all the rules have been triggered does pricing and
production processing take place.  In the past, each rule could trigger
its own processing.

ScaleFirmPriceSR, ScaleFirmProductionSR use getLastProduction() instead
of shortQuantityProduced.

With the above changes in place, it looks like networth values fall
within an acceptable range for the short Oligopoly experiment. 
Bankruptcies are rare, but this might be the result of the action system
not allowing for them to occur.

Consistency seems to be good.