Add endowment and prod cost to monopoly assay.
Some of the initial work on the AssayPriceProductionMonpoly class were
lost during the last merge.  I'm putting the code back in, and adding
the expansion of the assay to include endowment and production costs. 
Yes, some of the assay combinations are nonsensical (e.g. prod cost >
price), but I want a full assay to test the model's dynamics.